refactor(network): move to converters-protobuf consensusSignature to starknetapi BlockSignature
